chore(security): move firewall-middleware to firewall subdirectory and update imports
